pyproject.toml
[project]
name = "datasets"
version = "0.1.0"
description = "A simple dataset collection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.1",
    "httpx>=0.28.1",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "tqdm>=4.67.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
